fixed firstboot to only initialize /dev/sdb if it exists (only existing in vmware builds right now)
Test Plan: build core-image-tposs and start in qemu

Reviewers: tbreddin, hsychla

Reviewed By: tbreddin

CC: druss

Differential Revision: https://REVIEWSERVER/D964

+++ meta-tplino/recipes-core/firstboot/files/firstboot.sh
@@ -1,8 +1,11 @@
 #!/bin/sh
 
-(echo n; echo p; echo ; echo ; echo ; echo w;)|fdisk /dev/sdb
-mkfs.ext4 /dev/sdb1
-mkdir -p /srv/data
-echo "/dev/sdb1  /srv/data  ext4  defaults  0  2" >> /etc/fstab
+if [ -b /dev/sdb ]
+then
+    (echo n; echo p; echo ; echo ; echo ; echo w;)|fdisk /dev/sdb
+    mkfs.ext4 /dev/sdb1
+    mkdir -p /srv/data
+    echo "/dev/sdb1  /srv/data  ext4  defaults  0  2" >> /etc/fstab
+fi
 mount -a
 systemctl disable firstboot.service
